fix packing version print out
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@38043 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 608ab9ac49283057d92514ba708207a8267098f6

+++ LHCb/Event/PackedEvent/src/PackedMCCaloHit.cpp
@@ -1,4 +1,4 @@
-// $Id: PackedMCCaloHit.cpp,v 1.4 2009-11-07 12:20:27 jonrob Exp $
+// $Id: PackedMCCaloHit.cpp,v 1.5 2010-04-11 14:27:14 jonrob Exp $
 
 // local
 #include "Event/PackedMCCaloHit.h"
@@ -38,7 +38,7 @@ void MCCaloHitPacker::pack( const DataVector & hits,
   else
   {
     std::ostringstream mess;
-    mess << "Unknown packed data version " << phits.packingVersion();
+    mess << "Unknown packed data version " << (int)phits.packingVersion();
     throw GaudiException( mess.str(), "MCCaloHitPacker", StatusCode::FAILURE );
   }
 }
@@ -72,7 +72,7 @@ void MCCaloHitPacker::unpack( const PackedDataVector & phits,
   else
   {
     std::ostringstream mess;
-    mess << "Unknown packed data version " << phits.packingVersion();
+    mess << "Unknown packed data version " << (int)phits.packingVersion();
     throw GaudiException( mess.str(), "MCCaloHitPacker", StatusCode::FAILURE );
   }
 }
