+++ LHCb/Event/PackedEvent/src/PackedMCHit.cpp
@@ -1,4 +1,4 @@
-// $Id: PackedMCHit.cpp,v 1.4 2009-11-07 12:20:27 jonrob Exp $
+// $Id: PackedMCHit.cpp,v 1.5 2010-04-11 14:27:14 jonrob Exp $
 
 // STL
 #include <sstream>
@@ -48,7 +48,7 @@ void MCHitPacker::pack( const DataVector & hits,
   else
   {
     std::ostringstream mess;
-    mess << "Unknown packed data version " << phits.packingVersion();
+    mess << "Unknown packed data version " << (int)phits.packingVersion();
     throw GaudiException( mess.str(), "MCHitPacker", StatusCode::FAILURE );
   }
 }
@@ -89,7 +89,7 @@ void MCHitPacker::unpack( const PackedDataVector & phits,
   else
   {
     std::ostringstream mess;
-    mess << "Unknown packed data version " << phits.packingVersion();
+    mess << "Unknown packed data version " << (int)phits.packingVersion();
     throw GaudiException( mess.str(), "MCHitPacker", StatusCode::FAILURE );
   }
 }
