+++ LHCb/Event/PackedEvent/src/PackedMCRichDigitSummary.cpp
@@ -1,4 +1,4 @@
-// $Id: PackedMCRichDigitSummary.cpp,v 1.4 2009-11-07 12:20:27 jonrob Exp $
+// $Id: PackedMCRichDigitSummary.cpp,v 1.5 2010-04-11 14:27:14 jonrob Exp $
 
 // local
 #include "Event/PackedMCRichDigitSummary.h"
@@ -38,7 +38,7 @@ void MCRichDigitSummaryPacker::pack( const DataVector & sums,
   else
   {
     std::ostringstream mess;
-    mess << "Unknown packed data version " << psums.packingVersion();
+    mess << "Unknown packed data version " << (int)psums.packingVersion();
     throw GaudiException( mess.str(), "MCRichDigitSummaryPacker", StatusCode::FAILURE );
   }
 }
@@ -71,7 +71,7 @@ void MCRichDigitSummaryPacker::unpack( const PackedDataVector & psums,
   else
   {
     std::ostringstream mess;
-    mess << "Unknown packed data version " << psums.packingVersion();
+    mess << "Unknown packed data version " << (int)psums.packingVersion();
     throw GaudiException( mess.str(), "MCRichDigitSummaryPacker", StatusCode::FAILURE );
   }
 }
