+++ LHCb/Event/PackedEvent/src/PackedMCRichHit.cpp
@@ -1,4 +1,4 @@
-// $Id: PackedMCRichHit.cpp,v 1.4 2009-11-07 12:20:27 jonrob Exp $
+// $Id: PackedMCRichHit.cpp,v 1.5 2010-04-11 14:27:14 jonrob Exp $
 
 // local
 #include "Event/PackedMCRichHit.h"
@@ -41,7 +41,7 @@ void MCRichHitPacker::pack( const DataVector  & hits,
   else
   {
     std::ostringstream mess;
-    mess << "Unknown packed data version " << phits.packingVersion();
+    mess << "Unknown packed data version " << (int)phits.packingVersion();
     throw GaudiException( mess.str(), "MCRichHitPacker", StatusCode::FAILURE );
   }
 }
@@ -79,7 +79,7 @@ void MCRichHitPacker::unpack( const PackedDataVector & phits,
   else
   {
     std::ostringstream mess;
-    mess << "Unknown packed data version " << phits.packingVersion();
+    mess << "Unknown packed data version " << (int)phits.packingVersion();
     throw GaudiException( mess.str(), "MCRichHitPacker", StatusCode::FAILURE );
   }
 }
